add reset cluster button to workspace page
