+++ ui/src/app/app.module.ts
@@ -84,6 +84,7 @@ import {
 } from 'notebooks-generated';
 import { HighlightSearchComponent } from './highlight-search/highlight-search.component';
 import {ToolTipComponent} from './views/tooltip/component';
+import {ResetClusterButtonComponent} from "./views/reset-cluster-button/component";
 
 // Unfortunately stackdriver-errors-js doesn't properly declare dependencies, so
 // we need to explicitly load its StackTrace dep:
@@ -181,6 +182,7 @@ export function getLeoConfiguration(signInService: SignInService): LeoConfigurat
     WorkspaceWrapperComponent,
     HomepageComponent,
     HighlightSearchComponent,
+    ResetClusterButtonComponent,
   ],
   providers: [
     {
