+++ ui/src/app/views/workspace/component.html
@@ -169,6 +169,7 @@
     <a (click)="googleBucketModal = true">
       Google Bucket
     </a>
+    <react-cluster-button [billingProjectId]="workspace.namespace"></react-cluster-button>
   </div>
 </div>
 <app-workspace-share [workspace]="workspace"
